test: 스토리북 파일 페이지별로 분류
